create jpa repository for article model

+++ src/main/resources/db/migration/V2__create-topics.sql
@@ -1,6 +1,6 @@
 CREATE TABLE topics (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     description text NOT NULL,
-    user_id INTEGER,
+    user_id BIGINT,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
